+++ bika/lims/browser/worksheet/adapters/__init__.py
@@ -1,0 +1,2 @@
+# coding=utf-8
+from results import ResultOutOfRangeAdapter
