--- bika/lims/browser/worksheet/add_refsample.py
@@ -1,1 +1,0 @@
-# coding=utf-8
