+++ bika/lims/browser/worksheet/configure.zcml
@@ -4,11 +4,30 @@
     xmlns:i18n="http://namespaces.zope.org/i18n"
     i18n_domain="bika">
 
-    <!-- Worksheet Folder -->
+
+    <!-- Workflow actions -->
+    <browser:page
+      for="bika.lims.interfaces.IWorksheetFolder"
+      name="workflow_action"
+      class="bika.lims.browser.worksheet.workflow.WorksheetFolderWorkflowAction"
+      permission="zope.Public"
+      layer="bika.lims.interfaces.IBikaLIMS"
+    />
+
+    <browser:page
+      for="bika.lims.interfaces.IWorksheet"
+      name="workflow_action"
+      class="bika.lims.browser.worksheet.workflow.WorksheetWorkflowAction"
+      permission="zope.Public"
+      layer="bika.lims.interfaces.IBikaLIMS"
+    />
+
+
+    <!-- Views -->
     <browser:page
       for="bika.lims.interfaces.IWorksheetFolder"
       name="base_view"
-      class="bika.lims.browser.worksheet.worksheets.WorksheetsView"
+      class="bika.lims.browser.worksheet.views.FolderView"
       permission="zope.Public"
       layer="bika.lims.interfaces.IBikaLIMS"
     />
@@ -16,43 +35,15 @@
     <browser:page
       for="bika.lims.interfaces.IWorksheetFolder"
       name="worksheet_add"
-      class="bika.lims.browser.worksheetfolder.AddWorksheetView"
+      class="bika.lims.browser.worksheet.views.AddWorksheetView"
       permission="zope.Public"
       layer="bika.lims.interfaces.IBikaLIMS"
-    />
-
-    <!-- Workflow action button clicked -->
-    <browser:page
-      for="bika.lims.interfaces.IWorksheetFolder"
-      name="workflow_action"
-      class="bika.lims.browser.worksheetfolder.WorksheetFolderWorkflowAction"
-      permission="zope.Public"
-      layer="bika.lims.interfaces.IBikaLIMS"
-    />
-
-    <!-- Workflow action button clicked -->
-    <browser:page
-      for="bika.lims.interfaces.IWorksheet"
-      name="workflow_action"
-      class="bika.lims.browser.worksheet.WorksheetWorkflowAction"
-      permission="zope.Public"
-      layer="bika.lims.interfaces.IBikaLIMS"
-    />
-
-    <adapter
-      for="bika.lims.interfaces.IWorksheet"
-      factory="bika.lims.browser.worksheet.ResultOutOfRange"
-    />
-    <adapter
-      name="worksheet_priority_icons"
-      for="bika.lims.interfaces.IWorksheet"
-      factory="bika.lims.adapters.priorityicons.PriorityIcons"
     />
 
     <browser:page
       for="bika.lims.interfaces.IWorksheet"
       name="manage_results"
-      class="bika.lims.browser.worksheet.ManageResultsView"
+      class="bika.lims.browser.worksheet.views.ResultsView"
       permission="bika.lims.ViewResults"
       layer="bika.lims.interfaces.IBikaLIMS"
     />
@@ -60,7 +51,7 @@
     <browser:page
       for="bika.lims.interfaces.IWorksheet"
       name="add_analyses"
-      class="bika.lims.browser.worksheet.AddAnalysesView"
+      class="bika.lims.browser.worksheet.views.AddAnalysesView"
       permission="bika.lims.EditWorksheet"
       layer="bika.lims.interfaces.IBikaLIMS"
     />
@@ -68,7 +59,7 @@
     <browser:page
       for="bika.lims.interfaces.IWorksheet"
       name="add_blank"
-      class="bika.lims.browser.worksheet.AddBlankView"
+      class="bika.lims.browser.worksheet.views.AddBlankView"
       permission="bika.lims.EditWorksheet"
       layer="bika.lims.interfaces.IBikaLIMS"
     />
@@ -76,7 +67,7 @@
     <browser:page
       for="bika.lims.interfaces.IWorksheet"
       name="add_control"
-      class="bika.lims.browser.worksheet.AddControlView"
+      class="bika.lims.browser.worksheet.views.AddControlView"
       permission="bika.lims.EditWorksheet"
       layer="bika.lims.interfaces.IBikaLIMS"
     />
@@ -84,7 +75,7 @@
     <browser:page
       for="bika.lims.interfaces.IWorksheet"
       name="add_duplicate"
-      class="bika.lims.browser.worksheet.AddDuplicateView"
+      class="bika.lims.browser.worksheet.views.AddDuplicateView"
       permission="bika.lims.EditWorksheet"
       layer="bika.lims.interfaces.IBikaLIMS"
     />
@@ -92,23 +83,34 @@
     <browser:page
       for="bika.lims.interfaces.IWorksheet"
       name="export"
-      class="bika.lims.browser.worksheet.ExportView"
+      class="bika.lims.browser.worksheet.views.ExportView"
       permission="bika.lims.EditWorksheet"
       layer="bika.lims.interfaces.IBikaLIMS"
     />
 
     <browser:page
       for="bika.lims.interfaces.IWorksheet"
-      name="getServices"
-      class="bika.lims.browser.worksheet.ajaxGetServices"
+      name="getWorksheetReferences"
+      class="bika.lims.browser.worksheet.views.ReferenceSamplesView"
       permission="zope.Public"
       layer="bika.lims.interfaces.IBikaLIMS"
     />
 
     <browser:page
       for="bika.lims.interfaces.IWorksheet"
-      name="getWorksheetReferences"
-      class="bika.lims.browser.worksheet.ajaxGetWorksheetReferences"
+      name="print"
+      class="bika.lims.browser.worksheet.views.WorksheetPrintView"
+      permission="bika.lims.EditWorksheet"
+      layer="bika.lims.interfaces.IBikaLIMS"
+    />
+
+
+    <!-- Ajax signatures -->
+
+    <browser:page
+      for="bika.lims.interfaces.IWorksheet"
+      name="getServices"
+      class="bika.lims.browser.worksheet.ajax.GetServices"
       permission="zope.Public"
       layer="bika.lims.interfaces.IBikaLIMS"
     />
@@ -116,7 +118,7 @@
     <browser:page
       for="bika.lims.interfaces.IWorksheet"
       name="attachAnalyses"
-      class="bika.lims.browser.worksheet.ajaxAttachAnalyses"
+      class="bika.lims.browser.worksheet.ajax.AttachAnalyses"
       permission="zope.Public"
       layer="bika.lims.interfaces.IBikaLIMS"
     />
@@ -124,7 +126,7 @@
     <browser:page
       for="bika.lims.interfaces.IWorksheet"
       name="set_analyst"
-      class="bika.lims.browser.worksheet.ajaxSetAnalyst"
+      class="bika.lims.browser.worksheet.ajax.SetAnalyst"
       permission="zope.Public"
       layer="bika.lims.interfaces.IBikaLIMS"
     />
@@ -132,27 +134,22 @@
     <browser:page
       for="bika.lims.interfaces.IWorksheet"
       name="set_instrument"
-      class="bika.lims.browser.worksheet.ajaxSetInstrument"
+      class="bika.lims.browser.worksheet.ajax.SetInstrument"
       permission="zope.Public"
       layer="bika.lims.interfaces.IBikaLIMS"
     />
 
-    <!-- WS Print View -->
-    <browser:page
+
+    <!-- Adapters -->
+    <adapter
       for="bika.lims.interfaces.IWorksheet"
-      name="print"
-      class="bika.lims.browser.worksheet_print.WorksheetPrintView"
-      permission="bika.lims.EditWorksheet"
-      layer="bika.lims.interfaces.IBikaLIMS"
+      factory="bika.lims.browser.worksheet.adapters.ResultOutOfRangeAdapter"
     />
 
-    <!-- Multiple WS print preview -->
-    <browser:page
-      for="bika.lims.interfaces.IWorksheetFolder"
-      name="print"
-      class="bika.lims.browser.worksheet_print.WorksheetPrintView"
-      permission="bika.lims.EditWorksheet"
-      layer="bika.lims.interfaces.IBikaLIMS"
+    <adapter
+      name="worksheet_priority_icons"
+      for="bika.lims.interfaces.IWorksheet"
+      factory="bika.lims.adapters.priorityicons.PriorityIcons"
     />
 
 </configure>
