+++ bika/lims/browser/worksheet/views/__init__.py
@@ -1,0 +1,14 @@
+# coding=utf-8
+from add_analyses import AddAnalysesView
+from add_blank import AddBlankView
+from add_control import AddControlView
+from add_duplicate import AddDuplicateView
+from add_worksheet import AddWorksheetView
+from analyses import AnalysesView
+from analysisrequests import AnalysisRequestsView
+from export import ExportView
+from folder import FolderView
+from printview import PrintView
+from referencesamples import ReferenceSamplesView
+from results import ResultsView
+from services import ServicesView
