+++ bika/lims/browser/worksheet/views/add_blank.py
@@ -11,6 +11,8 @@ from bika.lims.browser.analyses import AnalysesView
 from bika.lims.browser.bika_listing import BikaListingView
 from bika.lims.browser.bika_listing import WorkflowAction
 from bika.lims.browser.referencesample import ReferenceSamplesView
+from bika.lims.browser.worksheet.tools import checkUserManage
+from bika.lims.browser.worksheet.tools import showRejectionMessage
 from bika.lims.exportimport import instruments
 from bika.lims.interfaces import IFieldIcons
 from bika.lims.interfaces import IWorksheet
@@ -62,7 +64,7 @@ class AddBlankView(BrowserView):
         if checkUserManage(self.context, self.request) == False:
             return []
 
-        rejected_alerts(self.context)
+        showRejectionMessage(self.context)
 
         form = self.request.form
         if 'submitted' in form:
